fix: [M3-5783] Feature Cleanup
